[executor] Support for multiple tasks
The executor now instantiates multiple gRPC clients and multiplexes
incoming mesos messages between them.
Each control message must now include the target descriptor, of which
for the executor the relevant part is the TaskID.
A custom common.TaskCommandInfo is now used instead of mesos.CommandInfo
to describe the task to be launched.
Arguments are not ignored any more if Shell == true, they are instead
appended to the call that's passed to the shell.

+++ executor/executorcmd/client.go
@@ -45,7 +45,7 @@ import (
 
 var log = logger.New(logrus.StandardLogger(), "executorcmd")
 
-func NewClient(controlPort uint16, controlMode controlmode.ControlMode) *RpcClient {
+func NewClient(controlPort uint64, controlMode controlmode.ControlMode) *RpcClient {
 	endpoint := fmt.Sprintf("127.0.0.1:%d", controlPort)
 	cxt, cancel := context.WithTimeout(context.Background(), 20*time.Second)
 	conn, err := grpc.DialContext(cxt, endpoint, grpc.WithInsecure())
